add feautures

+++ HillelPro_lesson13_part1/js/functions.js
@@ -1,3 +1,20 @@
+// Створити CRUD-додаток (Create, Read, Update, Delete):
+//
+// Виводиться список користувачів із кнопками “Edit”, “Remove”, “View” біля кожного користувача (use data-id attributes або event delegation) //done
+// список користувачів отримувати з js-файлу (масив об'єктів / використовувати функції-конструктори – за бажанням) //done
+//
+// При натисканні на кнопку “View” відкриваються дані користувача у блоці під списком //done
+// При натисканні на кнопку “Edit” з'являється можливість редагувати дані в блоці під списком. Дані зберігаються при натисканні на кнопку “Save” та оновлюють дані у списку
+// При натисканні на кнопку “Remove” користувач видаляється зі списку//done
+// Обов'язково підтвердження видалення (для уникнення видалення помилково)//done
+// Реалізувати можливість додавання нових користувачів//done
+// Бажано перевикористовувати форму редагування
+// При додаванні користувач з'являється у списку//done
+// Після перезавантаження сторінки всі зміни повинні зберігатись (використовувати localStorage)//done
+//
+
+
+
 function showRows(users) {
     for (let user of users) {
         showUserRow(user)
@@ -22,7 +39,7 @@ function showUserRow(user) {
             type: 'button',
             value: 'Edit',
             'data-type': 'edit',
-        }, {});
+        }, {click:editUserInformation});
 
     createElement(
         'input',
@@ -44,32 +61,53 @@ function showUserRow(user) {
             'data-type': 'view'
         }, {click: showUserCard})
 }
-
-function showAddUserForm() {
+function editUserInformation(event){
+   const userId = getUserId(event);
+   const user = getUserById(userId);
+   showAddUserForm(user);
+}
+function getUserById(userId) {
+    for (let user of users) {
+        if (user.id == userId) {
+            return user;
+        }
+    }
+}
+function showAddUserForm(user) {
     const parentSelector = '#form form';
+    const existingForm = document.querySelector(`#firstInput`);
 
-    createElement('input', parentSelector, '', {name: 'login', type: 'text', placeholder: 'Enter login',})
-    createElement('input', parentSelector, '', {name: 'name', type: 'text', placeholder: 'Enter name',})
-    createElement('input', parentSelector, '', {name: 'lastName', type: 'text', placeholder: 'Enter last name',})
-    createElement('input', parentSelector, '', {name: 'email', type: 'text', placeholder: 'Enter email',})
+    if (!existingForm) {
+        createElement('input', parentSelector, '', {name: 'login', type: 'text', placeholder: 'Enter login', id: 'firstInput',value:user.login || ''});
+        createElement('input', parentSelector, '', {name: 'name', type: 'text', placeholder: 'Enter name',value:user.name || ''});
+        createElement('input', parentSelector, '', {name: 'lastName', type: 'text', placeholder: 'Enter last name',value:user.lastName ||''});
+        createElement('input', parentSelector, '', {name: 'email', type: 'text', placeholder: 'Enter email',value:user.email||''});
 
-    createElement('input', parentSelector, '', {type: "button", value: 'Save'}, {click: handleSaveUser})
+        createElement('input', parentSelector, '', {type: "button", value: 'Save'}, {click: handleSaveUser});
+
+    }
+
 }
 
 function handleSaveUser() {
     const formElements = document.forms[0].elements;
 
+
     const login = formElements.login.value;
     const name = formElements.name.value;
     const lastName = formElements.lastName.value;
     const email = formElements.email.value;
+    const randomId = Date.now().toString().slice(-2);
+
 
     const user = {
+
+        id:randomId,   // id: this.id = users.length+1 получались дубли когда удалял элемент из середины следующий добавленый дублировал предыдущий.
         login,
         name,
         lastName,
         email,
-        id: this.id = users.length + 1,
+
     }
 
 
@@ -105,14 +143,33 @@ function updateStorage() {
 
 function handleDeleteUser(event) {
     const userId = getUserId(event)
-    deleteUserById(userId)
+    confirmation(userId)
+}
+
+function confirmation(userId) {
+    let confirmationBlock =document.getElementById('confirmationBlock');
+    if (!confirmationBlock) {
+        confirmationBlock = createElement('div', `div[data-user-id="${userId}"]`, `Are you sure you want to delete this user?`, {className: 'confirmation-block',id:'confirmationBlock'});
+        createElement('input', confirmationBlock, '', {type: 'button', value: 'YES'}, {
+            click: () => {
+                removeElement(confirmationBlock);
+                deleteUserById(userId);
+            }
+        })
+        createElement('input', confirmationBlock, '', {type: 'button', value: 'NO'}, {
+            click: () => {
+                removeElement(confirmationBlock);
+            }
+        })
+    }
+
 }
 
 function deleteUserById(id) {
     const indexToRemove = users.findIndex(user => user.id == id);
     users.splice(indexToRemove, 1);
 
-    removeElement(`div[data-user-id="${id}"]`)
+    removeElement(`div[data-user-id="${id}"]`);
 
     updateStorage();
 }
@@ -155,25 +212,33 @@ function showUserCard(event) {
     const container = document.getElementById('users');
     let userArticle = document.getElementById('userCard')
 
-    if(!userArticle){
-        userArticle = createElement('div', container, ``, {id: 'userCard','className':'user-item'});
+    if (!userArticle) {
+        userArticle = createElement('article', container, ``, {id: 'userCard', 'className': 'user-item'});
     }
 
-    displayInfoFromUser(event,userArticle)
+    displayInfoFromUser(event, userArticle)
 
 
 }
-function displayInfoFromUser(event,element){
-    const chosenUserId = getUserId(event)
-    for (let user of users){
-        if (user.id == chosenUserId){
-            element.innerHTML=`<h2>User INFO</h2>
-                                   <p>${user.login}</p>
-                                   <p>${user.name}</p>
-                                   <p>${user.lastName}</p>
-                                   <p>${user.email}</p> `
 
+function displayInfoFromUser(event, element) {
+    const chosenUserId = getUserId(event)
+    for (let user of users) {
+        if (user.id == chosenUserId) {
+            element.innerHTML = `<p>User INFO</p>
+                                <ul>
+                                  <li>User login: ${user.login}</li>
+                                  <li>User name: ${user.name}</li>
+                                  <li>User last name: ${user.lastName}</li>
+                                  <li>User email: ${user.email}</li>
+                               
+                                </ul>
+                               `
         }
 
     }
+    setTimeout(() => {
+        element.remove();
+    }, 3000);
+
 }
